Create fcfs_algorithm.py
We're going to implement the FCFS process scheduling algorithm to calculate the start and end times of all processes in the processes DataFrame. Recall that the FCFS algorithm schedules processes on the CPU as follows:

When a process arrives, if the processor is free, then assign the process to the processor. Otherwise add the process to a queue with all processes that are waiting for the processor.

Once a process finishes executing, take the process at the front of the queue and assign it to the CPU.

We'll calculate the start and end times of each process by simulating the FCFS algorithm. To do so, we'll keep track of the following data:

The current time. This is an integer that start at zero and increases by increments of one at each simulation step.
The total number of processes that finished executing. This value indicates when the simulation ends.
The waiting process queue. This is the queue that will store processes that arrive when the CPU is busy.
The ID of the current process running on the CPU. If the CPU is not running any process, this is set to None